Crate doubleclickbidmanager1

Features

Handle the following Resources with ease from the central hub ...

lineitems (downloadlineitems and uploadlineitems)

queries (createquery, deletequery, getquery, listqueries and runquery)

reports (listreports)

Everything else about the DoubleClick Bid Manager API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.queries().getquery(...).doit()
let r = hub.queries().createquery(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate doubleclickbidmanager1;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DoubleClickBidManager::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.queries().getquery("queryId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

DoubleClickBidManager
Central instance to access all DoubleClickBidManager related resource activities

DownloadLineItemsRequest
Request to fetch stored line items.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

downloadlineitems (request)

DownloadLineItemsResponse
Download line items response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

downloadlineitems (response)

FilterPair
Filter used to match traffic data in your report.

This type is not used in any activity, and only used as part of another schema.

LineitemDownloadlineitemMethodBuilder
Retrieves line items in CSV format.

LineitemMethodsBuilder
A builder providing access to all methods supported on lineitem resources. It is not used directly, but through the DoubleClickBidManager hub.

LineitemUploadlineitemMethodBuilder
Uploads line items in CSV format.

ListQueriesResponse
List queries response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listqueries (response)

ListReportsResponse
List reports response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listreports (response)

Parameters
Parameters of a query or report.

This type is not used in any activity, and only used as part of another schema.

Query
Represents a query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

getquery (response)

createquery (request|response)

QueryCreatequeryMethodBuilder
Creates a query.

QueryDeletequeryMethodBuilder
Deletes a stored query as well as the associated stored reports.

QueryGetqueryMethodBuilder
Retrieves a stored query.

QueryListqueryMethodBuilder
Retrieves stored queries.

QueryMetadata
Query metadata.

This type is not used in any activity, and only used as part of another schema.

QueryMethodsBuilder
A builder providing access to all methods supported on query resources. It is not used directly, but through the DoubleClickBidManager hub.

QueryRunqueryMethodBuilder
Runs a stored query to generate a report.

QuerySchedule
Information on how frequently and when to run a query.

This type is not used in any activity, and only used as part of another schema.

Report
Represents a report.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listreports (none)

ReportFailure
An explanation of a report failure.

This type is not used in any activity, and only used as part of another schema.

ReportKey
Key used to identify a report.

This type is not used in any activity, and only used as part of another schema.

ReportListreportMethodBuilder
Retrieves stored reports.

ReportMetadata
Report metadata.

This type is not used in any activity, and only used as part of another schema.

ReportMethodsBuilder
A builder providing access to all methods supported on report resources. It is not used directly, but through the DoubleClickBidManager hub.

ReportStatus
Report status.

This type is not used in any activity, and only used as part of another schema.

RowStatus
Represents the upload status of a row in the request.

This type is not used in any activity, and only used as part of another schema.

RunQueryRequest
Request to run a stored query to generate a report.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

runquery (request)

UploadLineItemsRequest
Request to upload line items.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

uploadlineitems (request)

UploadLineItemsResponse
Upload line items response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

uploadlineitems (response)

UploadStatus
Represents the status of upload.

This type is not used in any activity, and only used as part of another schema.